Updating the used Doxyfiles to version 1.9.5
Updating the used Doxyfiles to version 1.9.5

The following warnings are shown with `Doxyfile-man`:
```
.../waa.h:281: warning: unable to resolve reference to 'waa__find_common_base' for \ref command
.../waa.h:371: warning: unable to resolve reference to 'waa__find_common_base()' for \ref command
.../warnings.h:113: warning: unable to resolve reference to 'wa___warn_options' for \ref command

.../cache.h:50: warning: unable to resolve reference to 'cch__new_cache()' for \ref command
```
These are shown due to the fact that here, contrary to `Doxyfile` the setting of `EXTRACT_STATIC` is set to `NO` and thus the functions are not available.

```
.../cache.h:48: warning: unable to resolve reference to 'CACHE_DEFAULT' for \ref command
```
due to the fact that `EXTRACT_ALL` is set to `NO` and `CACHE_DEFAULT` is not documented, so no link possible.